Algorithms/0841.keys-and-rooms/KeysAndRooms.java
import java.util.LinkedList;
import java.util.List;
import java.util.Queue;

class KeysAndRooms {
    public boolean canVisitAllRooms(List<List<Integer>> rooms) {
        int n = rooms.size(), num = 0;
        boolean[] visited = new boolean[n];
        Queue<Integer> queue = new LinkedList<>();
        queue.offer(0);
        visited[0] = true;
        while (!queue.isEmpty()) {
            int roomNumber = queue.poll();
            num++;
            for (int key : rooms.get(roomNumber)) {
                if (!visited[key]) {
                    visited[key] = true;
                    queue.offer(key);
                }
            }
        }
        return num == n;
    }
}